Made background image responsive.
Made the background image responsive for better use with a tablet or phone. The original background image was interfering with a users ability to see the text on the web page. This was accomplished was by modifying the base.css file. The text for the main page was also changed to allow the text to be responsive.