Fixed the responsiveness for phones and tablets.
I needed to fix the responsive part of the base.css file. I had originally set it up with a max-width and this seemed to work fine with the browser but when it was testes on a phone it no longer worked like it should. Instead changing the code to have a min-width worked when tested on a phone.